# public/openapi.yaml
openapi: 3.0.0
info:
  title: API de Usuários e Posts
  version: 1.0.0
  description: API para gerenciar usuários e posts.
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      summary: Registra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: Usuário registrado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRegisterResponse"
        "400":
          description: Dados inválidos ou email já cadastrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    UserRegister:
      type: object
      properties:
        name:
          type: string
          description: Nome do usuário
          example: Maria Carvalho
        email:
          type: string
          format: email
          description: Endereço de e-mail do usuário
          example: maria.carvalho@exemple.com
        password:
          type: string
          minLength: 6
          description: Senha do usuário (mínimo 6 caracteres)
          example: minhasenha123
        role:
          type: string
          description: Papel/função do usuário
          example: "user"
          default: "user"
        image:
          type: string
          description: URL da imagem do usuário
          example: "https://example.com/avatar.jpg"
      required:
        - name
        - email
        - password
    UserRegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: "Usuário criado com sucesso"
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: 68fc0fd28318f76189c8dd91
            name:
              type: string
              example: João Silva
            email:
              type: string
              format: email
              example: maria.carvalho@example.com
            image:
              type: string
              example: null
            role:
              type: string
              example: "user"
            createdAt:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "Email já cadastrado"